[Chore] Rename `getEnvelopeListByEmail` to `getUserEnvelopes`
