+++ src/api/index.ts
@@ -1,6 +1,6 @@
 export { default as createEnvelope } from "./createEnvelope"
 export { default as createEnvelopeTemplate } from "./createEnvelopeTemplate"
 export { default as getEnvelopeById } from "./getEnvelopeById"
-export { default as getEnvelopeListByEmail } from "./getEnvelopeListByEmail"
 export { default as getTemplates } from "./getTemplates"
+export { default as getUserEnvelopes } from "./getUserEnvelopes"
 export { default as getUsersByQuery } from "./getUsersByQuery"
